added Pokemon class

+++ CIS17A-FInal/CIS17A-FInal/HoldItem.h
@@ -13,7 +13,7 @@ public:
 	HoldItem(double buffMultiplier);
 
 	int getHealAmount() { return _healAmount; }
-	int getBuffAmount() { return _buffMultiplier; }
+	double getBuffAmount() { return _buffMultiplier; }
 	//
 	~HoldItem();
 };
